+++ packages/gatsby/src/utils/worker/__tests__/queries.ts
@@ -20,6 +20,7 @@ import { getDataStore } from "../../../datastore"
 import { IGroupedQueryIds } from "../../../services"
 import { IGatsbyPage } from "../../../redux/types"
 import { runQueriesInWorkersQueue } from "../pool"
+import { readPageQueryResult } from "../../page-data"
 
 let worker: GatsbyTestWorkerPool | undefined
 
@@ -180,16 +181,11 @@ describeWhenLMDB(`worker (queries)`, () => {
     }
   })
 
-  // This was the original implementation of state syncing between a worker and the main process.
-  // We switched to "replaying actions" as a mechanism for state syncing.
-  // But we can get back to state saving / merging if "replaying actions" proves to be too expensive
-  // TODO: delete or re-activate depending on results yielded by "replaying actions" approach.
-  // The logic for `loadPartialStateFromDisk` itself is tested in `share-state` tests
   it(`should save worker "queries" state to disk`, async () => {
     if (!worker) fail(`worker not defined`)
 
     await worker.single.runQueries(queryIdsSmall)
-    await Promise.all(worker.all.saveQueries())
+    await Promise.all(worker.all.saveQueriesDependencies())
     // Pass "1" as workerId as the test only have one worker
     const result = loadPartialStateFromDisk([`queries`], `1`)
 
@@ -206,17 +202,7 @@ describeWhenLMDB(`worker (queries)`, () => {
           },
           "deletedQueries": Set {},
           "dirtyQueriesListToEmitViaWebsocket": Array [],
-          "queryNodes": Map {
-            "sq--q1" => Set {
-              "ceb8e742-a2ce-5110-a560-94c93d1c71a5",
-            },
-            "/foo" => Set {
-              "ceb8e742-a2ce-5110-a560-94c93d1c71a5",
-            },
-            "/bar" => Set {
-              "ceb8e742-a2ce-5110-a560-94c93d1c71a5",
-            },
-          },
+          "queryNodes": Map {},
           "trackedComponents": Map {},
           "trackedQueries": Map {
             "sq--q1" => Object {
@@ -262,8 +248,9 @@ describeWhenLMDB(`worker (queries)`, () => {
     await worker.single.runQueries(queryIdsSmall)
     const stateFromWorker = await worker.single.getState()
 
-    const pageQueryResult = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_foo.json`
+    const pageQueryResult = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/foo`
     )
 
     expect(pageQueryResult.data).toStrictEqual({
@@ -279,8 +266,9 @@ describeWhenLMDB(`worker (queries)`, () => {
     await worker.single.runQueries(queryIdsSmall)
     const stateFromWorker = await worker.single.getState()
 
-    const pageQueryResult = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_bar.json`
+    const pageQueryResult = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/bar`
     )
 
     expect(pageQueryResult.data).toStrictEqual({
@@ -301,8 +289,9 @@ describeWhenLMDB(`worker (queries)`, () => {
     const stateFromWorker = await worker.single.getState()
 
     // Called the complete ABC so we can test _a
-    const pageQueryResultA = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_a.json`
+    const pageQueryResultA = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/a`
     )
 
     expect(pageQueryResultA.data).toStrictEqual({
@@ -311,8 +300,9 @@ describeWhenLMDB(`worker (queries)`, () => {
       },
     })
 
-    const pageQueryResultZ = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_z.json`
+    const pageQueryResultZ = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/z`
     )
 
     expect(pageQueryResultZ.data).toStrictEqual({
@@ -350,10 +340,9 @@ describeWhenLMDB(`worker (queries)`, () => {
     const expectedActionShapes = {
       QUERY_START: [`componentPath`, `isPage`, `path`],
       PAGE_QUERY_RUN: [`componentPath`, `isPage`, `path`, `resultHash`],
-      CREATE_COMPONENT_DEPENDENCY: [`nodeId`, `path`],
       ADD_PENDING_PAGE_DATA_WRITE: [`path`],
     }
-    expect(result).toBeArrayOfSize(11)
+    expect(result).toBeArrayOfSize(8)
 
     for (const action of result) {
       expect(action.type).toBeOneOf(Object.keys(expectedActionShapes))
@@ -380,14 +369,6 @@ describeWhenLMDB(`worker (queries)`, () => {
       },
       {
         payload: {
-          nodeId: `ceb8e742-a2ce-5110-a560-94c93d1c71a5`,
-          path: `sq--q1`,
-        },
-        plugin: ``,
-        type: `CREATE_COMPONENT_DEPENDENCY`,
-      },
-      {
-        payload: {
           componentPath: `/static-query-component.js`,
           isPage: false,
           path: `sq--q1`,
@@ -411,22 +392,6 @@ describeWhenLMDB(`worker (queries)`, () => {
           path: `/bar`,
         },
         type: `QUERY_START`,
-      },
-      {
-        payload: {
-          nodeId: `ceb8e742-a2ce-5110-a560-94c93d1c71a5`,
-          path: `/foo`,
-        },
-        plugin: ``,
-        type: `CREATE_COMPONENT_DEPENDENCY`,
-      },
-      {
-        payload: {
-          nodeId: `ceb8e742-a2ce-5110-a560-94c93d1c71a5`,
-          path: `/bar`,
-        },
-        plugin: ``,
-        type: `CREATE_COMPONENT_DEPENDENCY`,
       },
       {
         payload: {