+++ packages/gatsby-admin/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.21.0-next.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.21.0-next.5...gatsby-admin@0.21.0-next.6) (2021-07-27)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.21.0-next.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.21.0-next.4...gatsby-admin@0.21.0-next.5) (2021-07-26)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.21.0-next.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.21.0-next.3...gatsby-admin@0.21.0-next.4) (2021-07-23)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.21.0-next.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.21.0-next.2...gatsby-admin@0.21.0-next.3) (2021-07-23)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.21.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.21.0-next.1...gatsby-admin@0.21.0-next.2) (2021-07-22)
+
+**Note:** Version bump only for package gatsby-admin
+
 # [0.21.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.21.0-next.0...gatsby-admin@0.21.0-next.1) (2021-07-19)
 
 **Note:** Version bump only for package gatsby-admin